+++ templates/product.single.js
@@ -10,6 +10,7 @@
         // Copy, format and return the config and products
         options.template_name = 'okanjo-product-single';
         options.config = data.config;
+        options.proxy_url = this.proxy_url;
         options.products = okanjo.mvc.formats.product(data.products);
         return options;
     }, {